sort shared functions
