+++ utils.py
@@ -6,6 +6,8 @@ import logging
 import json
 
 
+
+#This is Shared 
 class MissingFieldException(Exception):
     """
     Exception for a missing field in a dict
@@ -13,6 +15,7 @@ class MissingFieldException(Exception):
     pass
 
 
+#This is shared 
 @dataclass
 class GalaxyDataset:
     """
@@ -45,6 +48,7 @@ class GalaxyDataset:
         self.reference_name = info["metadata_dbkey"]
 
 
+## This is Shared
 def set_up_galaxy_instance(galaxy_url: str, galaxy_key: str) -> GalaxyInstance:
     """
     Returns a galaxy instance with the given URL and api key.
@@ -94,7 +98,7 @@ def set_up_galaxy_instance(galaxy_url: str, galaxy_key: str) -> GalaxyInstance:
     # resp = gi.make_get_request(galaxy_url + "/api/configuration?keys=allow_user_deletion").content
     return gi
 
-
+#This is Shared
 def string_as_bool(string: Any) -> bool:
     """
     Returns the bool value corresponding to the given value
@@ -121,8 +125,7 @@ def get_beacon_histories(gi: GalaxyInstance) -> List[str]:
 
     # get histories from galaxy api
     # URL is used because the name filter is not supported by bioblend as of now
-    response: Response = gi.make_get_request(
-        f"{gi.base_url}/api/histories?q=name&qv=Beacon%20Export%20%F0%9F%93%A1&all=true&deleted=false")
+    response: Response = gi.make_get_request(f"{gi.base_url}/api/histories?q=name&qv=Beacon%20Export%20%F0%9F%93%A1&all=true&deleted=false")
 
     # check if the reuest was successful
     if response.status_code != 200:
@@ -140,101 +143,9 @@ def get_beacon_histories(gi: GalaxyInstance) -> List[str]:
 
         # skip adding the history if beacon_enabled is not set for the owner account
         history_user_preferences: Dict[str, str] = user_details["preferences"]
-        if "beacon_enabled" not in history_user_preferences or not string_as_bool(
-                history_user_preferences["beacon_enabled"]):
+        if "beacon_enabled" not in history_user_preferences or not string_as_bool(history_user_preferences["beacon_enabled"]):
             continue
 
         history_ids.append(history["id"])
 
     return history_ids
-
-
-def get_datasets(gi: GalaxyInstance, history_id: str) -> List[GalaxyDataset]:
-    """
-    Fetches a given histories datasets from galaxy
-
-        Parameters:
-            gi (GalaxyInstance): galaxy instance to be used for the request
-            history_id (str): (encoded) ID of the galaxy history
-
-        Returns:
-            datasets (List[GalaxyDataset]): list of all datasets in the given history
-    """
-
-    # datasets = gi.histories.show_matching_datasets(history_id)
-
-    datasets: List[GalaxyDataset] = []
-
-    offset: int = 0
-    limit: int = 500
-
-    # galaxy api uses paging for datasets. This while loop continuously retrieves pages of *limit* datasets
-    # The loop breaks the first time an empty page comes up
-    while True:
-        # TODO extensions only allow one entry atm
-        # retrieve a list of datasets from the galaxy api
-        api_dataset_list = gi.datasets.get_datasets(
-            history_id=history_id,
-            deleted=False,
-            extension=["json", "json_bgzip"],
-            limit=limit,
-            offset=offset
-        )
-
-        # each api_dataset_list_entry is a dictionary with the fields:
-        #    "id", "name", "history_id", "hid", "history_content_type", "deleted", "visible",
-        #    "type_id", "type", "create_time", "update_time", "url", "tags", "dataset_id",
-        #    "state", "extension", "purged"
-        for api_dataset_list_entry in api_dataset_list:
-            dataset_info = gi.datasets.show_dataset(dataset_id=api_dataset_list_entry["id"])
-            # read dataset information from api
-            try:
-                dataset = GalaxyDataset(dataset_info)
-            except MissingFieldException as e:
-                # the exception is thrown by the constructor of GalaxyDataset which checks if all keys that are used
-                # actually exists
-                logging.warning(
-                    f"not reading dataset {api_dataset_list_entry['id']} because {e} from api response")
-
-            # filter for valid human references
-            # match = re.match(r"(GRCh\d+|hg\d+).*", dataset.reference_name)
-            # if match is None:
-            #     # skip datasets with unknown references
-            #     logging.warning(
-            #         f"not reading dataset {dataset.name} with unknown reference \"{dataset.reference_name}\"")
-            #     continue
-
-            # set reference name to the first match group
-            #
-            # THIS WILL REMOVE PATCH LEVEL FROM THE REFERENCE
-            # therefore all patch levels will be grouped under the major version of the reference
-            # dataset.reference_name = match.group(1)
-
-            datasets.append(dataset)
-        offset += limit
-
-        # no entries left
-        if len(api_dataset_list) == 0:
-            break
-
-    # return the finished dataset list
-    return datasets
-
-def download_dataset(gi: GalaxyInstance, dataset: GalaxyDataset, filename: str) -> None:
-    """
-    Downloads a dataset from galaxy to a given path
-
-        Parameters:
-            gi (GalaxyInstance): galaxy instance to download from
-            dataset (GalaxyDataset): the dataset to download
-            filename (str): output filename including complete path
-
-        Returns:
-            Nothing
-
-    """
-    try:
-        gi.datasets.download_dataset(dataset.id, filename, use_default_filename=False)
-    except Exception as e:
-        # TODO catch exceptions
-        logging.critical(f"something went wrong while downloading file - {e} filename:{filename}")
